test_db.py
import mysql.connector
import os
from dotenv import load_dotenv

load_dotenv()

try:
    connection = mysql.connector.connect(
        host=os.getenv('MYSQL_HOST', 'localhost'),
        user=os.getenv('MYSQL_USER', 'root'),
        password=os.getenv('MYSQL_PASSWORD', '')
    )
    if connection.is_connected():
        print("Successfully connected to MySQL server.")
        connection.close()
except Exception as e:
    print(f"Error: {e}")
