requirements.txt
flask
flask-cors
mysql-connector-python
pandas
numpy
scikit-learn
joblib
python-dotenv
matplotlib
seaborn
flask-sqlalchemy
simplejson
